backend/requirements.txt
Flask==2.1.2
Werkzeug==2.1.2
python-dotenv==0.19.0
openai==0.27.0
tiktoken==0.7.0
flask-cors==3.0.10
